✨ Convert to Express Server Application and Expose Web Interface
