+++ package.json
@@ -24,6 +24,9 @@
   },
   "homepage": "https://github.com/alexlee-dev/app-color-organizer#readme",
   "devDependencies": {
+    "@types/cors": "^2.8.6",
+    "@types/express": "^4.17.3",
+    "@types/morgan": "^1.9.0",
     "@types/node": "^13.9.8",
     "@types/node-fetch": "^2.5.5",
     "@types/sharp": "^0.24.0",
@@ -39,6 +42,9 @@
     "app-store-scraper": "^0.16.2",
     "chalk": "^3.0.0",
     "colorthief": "^2.3.0",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "morgan": "^1.10.0",
     "node-fetch": "^2.6.0",
     "ora": "^4.0.3",
     "pickitt": "^3.2.0",
