+++ src/index.ts
@@ -1,77 +1,44 @@
-/* eslint-disable object-curly-newline */
-/* eslint-disable no-await-in-loop */
-/* eslint-disable no-plusplus */
-/* eslint-disable no-unused-vars */
 /* eslint-disable import/extensions */
-import ColorThief from 'colorthief';
-import { promises } from 'fs';
-import fetch from 'node-fetch';
-import ora from 'ora';
+/* eslint-disable no-console */
+import cors from 'cors';
+import express from 'express';
+import morgan from 'morgan';
 import path from 'path';
-import { readIt } from 'pickitt';
-import { getArtworkUrl, saveImage, rgb2Hsl, sortByHue } from './util';
-import { App, AppData } from './types';
 
-const main = async (): Promise<void> => {
-  let currentSpinner = null;
-  try {
-    const readingDataSpinner = ora('Reading Data').start();
-    currentSpinner = readingDataSpinner;
-    const apps: App[] = await readIt(path.join(__dirname, '../data/apps.json'));
-    readingDataSpinner.succeed('Successfully Read Data');
+import colorange from './colorange';
 
-    const data: AppData[] = [];
+const app = express();
+const port = 3000;
 
-    const gettingArtworkSpinner = ora('Getting Artwork URLs').start();
-    currentSpinner = gettingArtworkSpinner;
-    for (let i = 0; i < apps.length; i++) {
-      const app: App = apps[i];
-      const icon = await getArtworkUrl(app.name);
-      data.push({ colors: null, icon, name: app.name });
+app.use(express.json());
+
+app.use(morgan('dev'));
+
+const whitelistDomains = ['http://localhost:3000', undefined];
+
+const corsOptions = {
+  origin: (origin: string, cb: Function): void => {
+    if (whitelistDomains.indexOf(origin) !== -1) {
+      cb(null, true);
+    } else {
+      console.log(`Colorange Sever refused to allow: ${origin}`);
+      cb(new Error('Not allowed by CORS'));
     }
-    gettingArtworkSpinner.succeed('Successfully Retrieved Artwork URLs');
-
-    const convertingImagesSpinner = ora('Converting Images').start();
-    currentSpinner = convertingImagesSpinner;
-    for (let i = 0; i < apps.length; i++) {
-      const response = await fetch(data[i].icon);
-      const buffer = await response.buffer();
-      await saveImage(buffer, `icon${i}.jpg`);
-
-      const colors = await ColorThief.getColor(
-        path.join(__dirname, `../icon${i}.jpg`),
-      );
-
-      data[i].colors = colors;
-    }
-    convertingImagesSpinner.succeed('Successfully Converted Images');
-
-    const cleaningFilesSpinner = ora('Cleaning Files').start();
-    currentSpinner = cleaningFilesSpinner;
-    for (let i = 0; i < apps.length; i++) {
-      await promises.unlink(path.join(__dirname, `../icon${i}.jpg`));
-    }
-    cleaningFilesSpinner.succeed('Successfully Cleaned Files');
-
-    const sortingApplicationsSpinner = ora('Sorting Applications').start();
-    currentSpinner = sortingApplicationsSpinner;
-    // * Convert colors from RGB to HSL
-    data.forEach((app: AppData, i: number) => {
-      data[i].colors = rgb2Hsl(app.colors);
-    });
-
-    const sortedData = sortByHue(data);
-
-    const sortedNames = [];
-
-    sortedData.forEach((app: AppData) => sortedNames.push(app.name));
-
-    sortingApplicationsSpinner.succeed('Successfully Sorted Applications');
-    console.log(sortedNames);
-  } catch (e) {
-    currentSpinner.fail();
-    console.error(e);
-  }
+  },
 };
 
-main();
+app.use(cors(corsOptions));
+
+app.use(express.static(path.join(__dirname, '../public')));
+
+app.post('/test', async (req, res) => {
+  const { apps } = req.body;
+  const sortedColors: string[] = await colorange(apps);
+  res.send(sortedColors);
+});
+
+app.get('*', (req, res) => {
+  res.status(404).send();
+});
+
+app.listen(port, () => console.log(`Server is listening on port: ${port}`));
